ENH+RF: Rename skl dataset wrappers for consistency (fixes gh-118)
Full resolution is pending scikit-learn/scikit-learn#2865
